Testing add_token linked list with while loop.
Try this one out, I made it so that if nothing is in the linked list, it will add a new token and if there are, I used a while loop to iterate to the end of the list and it should add the new token to the end of the list. However, I think there is a better way to do it if every time add_token is called, it adds the new token to the end of the list without iterating it each time.